fix x64 build warnings

+++ src/priv/util/engine/resourceLRU.cpp
@@ -10,7 +10,7 @@
 //
 //
 // -----------------------------------------------------------------------------
-bool GN::engine::RenderEngine::ResourceLRU::init( UInt32 maxTexBytes, UInt32 maxMeshBytes )
+bool GN::engine::RenderEngine::ResourceLRU::init( size_t maxTexBytes, size_t maxMeshBytes )
 {
     GN_GUARD;
